requirements.txt
# Requirements for the Call Tracking sample Django project

# Django (1.8 is a long-term support release)
Django==1.8

# Third-party app dependencies
django-bootstrap3==6.1.0
django-forms-bootstrap==3.0.1
django-phonenumber-field==0.7.2
phonenumbers==7.0.8
psycopg2==2.6.1
twilio==4.4.0
whitenoise==2.0.2

# Production dependencies
dj-database-url==0.3.0
gunicorn==19.3.0

# Test dependencies
coverage==3.7.1
mock==1.3.0
model-mommy==1.2.5
six==1.9.0
